add compile commands
