+++ docs/COMMANDS.md
@@ -2,15 +2,17 @@
 
 Longform provides a number of commands you may use to quickly navigate and edit your projects. Those marked as “requires editing” below require you to be actively editing a scene in your project.
 
-| Name                             | Requires Editing | Description                                                                                                                             |
-| -------------------------------- | ---------------- | --------------------------------------------------------------------------------------------------------------------------------------- |
-| Open Current Note’s Project      | true             | Selects your current project (or draft of project) in the Longform pane.                                                                |
-| Previous Scene                   | true             | Opens the previous scene to the one you’re currently editing, if available.                                                             |
-| Previous Scene at Indent         | true             | Opens the previous scene to the one you’re currently editing at the same indentation level, if available.                               |
-| Next Scene                       | true             | Opens the next scene to the one you’re currently editing, if available.                                                                 |
-| Next Scene at Indent             | true             | Opens the next scene to the one you’re currently editing at the same indentation level, if available.                                   |
-| Indent Scene                     | true             | Increases the current scene indentation level by one.                                                                                   |
-| Unindent Scene                   | true             | Decreases the current scene indentation level by one.                                                                                   |
-| Jump to Project                  | false            | Fuzzy-matches against known project titles, then shows it in the Longform pane and switches to that pane.                               |
-| Jump to Scene in Current Project | false            | Shows all scenes in the current project (in order and indented), and allows fuzzy-matching against them to quickly switch to a new one. |
-| Open Longform Pane               | false            | Switches to the Longform pane.                                                                                                          |
+| Name                                          | Requires Editing | Description                                                                                                                             |
+| --------------------------------------------- | ---------------- | --------------------------------------------------------------------------------------------------------------------------------------- |
+| Open Current Note’s Project                   | true             | Selects your current project (or draft of project) in the Longform pane.                                                                |
+| Previous Scene                                | true             | Opens the previous scene to the one you’re currently editing, if available.                                                             |
+| Previous Scene at Indent                      | true             | Opens the previous scene to the one you’re currently editing at the same indentation level, if available.                               |
+| Next Scene                                    | true             | Opens the next scene to the one you’re currently editing, if available.                                                                 |
+| Next Scene at Indent                          | true             | Opens the next scene to the one you’re currently editing at the same indentation level, if available.                                   |
+| Indent Scene                                  | true             | Increases the current scene indentation level by one.                                                                                   |
+| Unindent Scene                                | true             | Decreases the current scene indentation level by one.                                                                                   |
+| Jump to Project                               | false            | Fuzzy-matches against known project titles, then shows it in the Longform pane and switches to that pane.                               |
+| Jump to Scene in Current Project              | false            | Shows all scenes in the current project (in order and indented), and allows fuzzy-matching against them to quickly switch to a new one. |
+| Open Longform Pane                            | false            | Switches to the Longform pane.                                                                                                          |
+| Compile current project with current workflow | false            | Compiles whatever project is currently selected with its currently-selected workflow.                                                   |
+| Compile project…                              | false            | Select a project, draft, and workflow, then compile using those selections.                                                             |
