+++ src/commands/navigation.ts
@@ -1,10 +1,5 @@
-import {
-  FuzzySuggestModal,
-  Keymap,
-  type App,
-  type Instruction,
-  type PaneType,
-} from "obsidian";
+import type { App, PaneType } from "obsidian";
+
 import { get } from "svelte/store";
 import { repeat } from "lodash";
 
@@ -26,6 +21,7 @@ import { VIEW_TYPE_LONGFORM_EXPLORER } from "src/view/explorer/ExplorerPane";
 import type LongformPlugin from "src/main";
 import type { Draft } from "src/model/types";
 import { draftTitle } from "src/model/draft-utils";
+import { JumpModal } from "./helpers";
 
 const checkForLocation = (
   checking: boolean,
@@ -143,49 +139,6 @@ export const showLongform: CommandBuilder = (plugin) => ({
     showLeaf(plugin);
   },
 });
-
-class JumpModal<T> extends FuzzySuggestModal<string> {
-  items: Map<string, T>;
-  onSelect: (value: T, modEvent: boolean | PaneType) => void;
-
-  constructor(
-    app: App,
-    items: Map<string, T>,
-    instructions: Instruction[] = [],
-    onSelect: (value: T, modEvent: boolean | PaneType) => void
-  ) {
-    super(app);
-
-    this.items = items;
-    this.onSelect = onSelect;
-
-    this.scope.register(["Meta"], "Enter", (evt) => {
-      const result = this.containerEl.getElementsByClassName(
-        "suggestion-item is-selected"
-      );
-      if (result.length > 0) {
-        const selected = result[0].innerHTML;
-        this.onChooseItem(selected, evt);
-      }
-      this.close();
-      return false;
-    });
-
-    this.setInstructions(instructions);
-  }
-
-  getItems(): string[] {
-    return Array.from(this.items.keys());
-  }
-
-  getItemText(item: string): string {
-    return item;
-  }
-
-  onChooseItem(item: string, evt: MouseEvent | KeyboardEvent): void {
-    this.onSelect(this.items.get(item), Keymap.isModEvent(evt));
-  }
-}
 
 export const jumpToProject: CommandBuilder = (plugin) => ({
   id: "longform-jump-to-project",
