+++ test-longform-vault/.obsidian/core-plugins-migration.json
@@ -25,5 +25,6 @@
   "file-recovery": true,
   "publish": false,
   "sync": true,
-  "canvas": true
+  "canvas": true,
+  "bookmarks": true
 }
