+++ test-longform-vault/.obsidian/core-plugins.json
@@ -10,6 +10,7 @@
   "note-composer",
   "command-palette",
   "editor-status",
+  "bookmarks",
   "markdown-importer",
   "word-count",
   "file-recovery",
