+++ test-longform-vault/projects 2.0/A Novel Draft 2/manuscript.md
@@ -1,0 +1,21 @@
+# in this draft
+
+asdasdsadasd asdsa asd 
+
+---
+
+# test
+
+
+
+---
+
+# test3
+
+sadasdsadsad
+
+---
+
+# test4
+
+asdsadasdasd
