+++ test-longform-vault/projects 2.0/A Short Story.md
@@ -1,7 +1,7 @@
 ---
 longform:
   format: single
-  workflow: single-scene flow2
+  workflow: Default Workflow
 ---
 
 This is a short story, and an example of a single-note Longform project. The entire project is this note. It should display differently from multi-note, more traditional Longform projects.
